rename .api.md files and csv
