--- common/api/ExpressServer.api.md
@@ -1,44 +1,0 @@
-## API Report File for "@itwin/express-server"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { Application } from 'express';
-import { RpcConfiguration } from '@itwin/core-common';
-import { Server } from 'http';
-import { WebAppRpcProtocol } from '@itwin/core-common';
-
-// @public
-export class IModelJsExpressServer {
-    constructor(protocol: WebAppRpcProtocol, config?: IModelJsExpressServerConfig);
-    // (undocumented)
-    protected _app: Application;
-    // (undocumented)
-    protected _configureHeaders(): void;
-    // (undocumented)
-    protected _configureMiddleware(): void;
-    // (undocumented)
-    protected _configureRoutes(): void;
-    static readonly defaults: IModelJsExpressServerConfig;
-    initialize(port: number | string): Promise<Server>;
-    // @alpha (undocumented)
-    get rpcConfiguration(): RpcConfiguration;
-}
-
-// @public
-export interface IModelJsExpressServerConfig {
-    // (undocumented)
-    uploadLimit: string;
-}
-
-// @alpha (undocumented)
-export class WebEditServer extends IModelJsExpressServer {
-    constructor(protocol: WebAppRpcProtocol, config?: IModelJsExpressServerConfig);
-    // (undocumented)
-    protected _configureRoutes(): void;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
