--- common/api/Reporter.api.md
@@ -1,16 +1,0 @@
-## API Report File for "@itwin/perf-tools"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-// @beta (undocumented)
-export class Reporter {
-    addEntry(testSuite: string, testName: string, valueDescription: string, value: number, info?: any, date?: string): void;
-    clearEntries(): void;
-    exportCSV(fileName: string): void;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
