--- common/api/__DOC_ONLY_ELECTRON__.api.md
@@ -1,103 +1,0 @@
-## API Report File for "@itwin/core-electron"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { AsyncMethodsOf } from '@itwin/core-bentley';
-import type { BrowserWindow } from 'electron';
-import type { BrowserWindowConstructorOptions } from 'electron';
-import type * as ElectronModule from 'electron';
-import { IpcHandler } from '@itwin/core-backend';
-import { NativeAppOpts } from '@itwin/core-frontend';
-import { NativeHostOpts } from '@itwin/core-backend';
-import { PickAsyncMethods } from '@itwin/core-bentley';
-import { RpcConfiguration } from '@itwin/core-common';
-import { RpcInterfaceDefinition } from '@itwin/core-common';
-import type { WebPreferences } from 'electron';
-
-// @beta @deprecated
-export type DialogModuleMethod = AsyncMethodsOf<Electron.Dialog>;
-
-// @beta
-export class ElectronApp {
-    static dialogIpc: PickAsyncMethods<Electron.Dialog>;
-    // (undocumented)
-    static get isValid(): boolean;
-    // (undocumented)
-    static shutdown(): Promise<void>;
-    static startup(opts?: ElectronAppOpts): Promise<void>;
-}
-
-// @beta (undocumented)
-export type ElectronAppOpts = NativeAppOpts;
-
-// @beta
-export class ElectronHost {
-    // (undocumented)
-    static get app(): Electron.App;
-    // (undocumented)
-    static appIconPath: string;
-    // (undocumented)
-    static get electron(): typeof ElectronModule;
-    // (undocumented)
-    static frontendURL: string;
-    static getWindowMaximizedSetting(windowName: string): boolean | undefined;
-    static getWindowSizeAndPositionSetting(windowName: string): WindowSizeAndPositionProps | undefined;
-    // (undocumented)
-    static get ipcMain(): Electron.IpcMain;
-    // (undocumented)
-    static get isValid(): boolean;
-    static get mainWindow(): BrowserWindow | undefined;
-    static openMainWindow(windowOptions?: ElectronHostWindowOptions): Promise<void>;
-    // (undocumented)
-    static rpcConfig: RpcConfiguration;
-    static startup(opts?: ElectronHostOpts): Promise<void>;
-    // (undocumented)
-    static webResourcesPath: string;
-}
-
-// @beta
-export interface ElectronHostOptions {
-    developmentServer?: boolean;
-    frontendPort?: number;
-    frontendURL?: string;
-    iconName?: string;
-    ipcHandlers?: (typeof IpcHandler)[];
-    rpcInterfaces?: RpcInterfaceDefinition[];
-    webResourcesPath?: string;
-}
-
-// @beta (undocumented)
-export interface ElectronHostOpts extends NativeHostOpts {
-    // (undocumented)
-    electronHost?: ElectronHostOptions;
-}
-
-// @beta (undocumented)
-export interface ElectronHostWindowOptions extends BrowserWindowConstructorOptions {
-    storeWindowName?: string;
-    titleBarStyle?: ("default" | "hidden" | "hiddenInset" | "customButtonsOnHover");
-    webPreferences?: Omit<WebPreferences, "preload" | "experimentalFeatures" | "nodeIntegration" | "contextIsolation" | "sandbox" | "nodeIntegrationInWorker" | "nodeIntegrationInSubFrames">;
-}
-
-// @internal (undocumented)
-export const electronIpcStrings: {
-    readonly dialogChannel: "itwinjs-electron/dialog";
-};
-
-// @beta
-export interface WindowSizeAndPositionProps {
-    // (undocumented)
-    height: number;
-    // (undocumented)
-    width: number;
-    // (undocumented)
-    x: number;
-    // (undocumented)
-    y: number;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
