+++ core/electron/package.json
@@ -11,8 +11,8 @@
     "build:cjs": "tsc 1>&2 --outDir lib/cjs",
     "copy:test-assets": "cpx \"./src/test/assets/**/*\" ./lib/cjs/test/assets",
     "clean": "rimraf -g lib .rush/temp/package-deps*.json",
-    "docs": "betools docs --json=../../generated-docs/core/core-electron/file.json --tsIndexFile=./__DOC_ONLY_ELECTRON__.ts --onlyJson",
-    "extract-api": "betools extract-api --entry=__DOC_ONLY_ELECTRON__",
+    "docs": "betools docs --json=../../generated-docs/core/core-electron/file.json --tsIndexFile=./core-electron.ts --onlyJson",
+    "extract-api": "betools extract-api --entry=core-electron",
     "lint": "eslint \"./src/**/*.ts\" 1>&2",
     "test": "",
     "test:integration": "npm run test:integration:backend && npm run test:integration:frontend",
