+++ core/express-server/package.json
@@ -2,9 +2,9 @@
   "name": "@itwin/express-server",
   "version": "5.0.0-dev.108",
   "description": "iTwin.js express utilities",
-  "main": "lib/cjs/ExpressServer.js",
-  "module": "lib/esm/ExpressServer.js",
-  "typings": "lib/cjs/ExpressServer",
+  "main": "lib/cjs/express-server.js",
+  "module": "lib/esm/express-server.js",
+  "typings": "lib/cjs/express-server",
   "license": "MIT",
   "engines": {
     "node": "^20.0.0 || ^22.0.0"
@@ -14,8 +14,8 @@
     "build:cjs": "tsc 1>&2 --outDir lib/cjs",
     "build:esm": "tsc 1>&2 --module ES2022 --outDir lib/esm",
     "clean": "rimraf -g lib .rush/temp/package-deps*.json",
-    "docs": "betools docs --json=../../generated-docs/core/express-server/file.json --tsIndexFile=./ExpressServer.ts --onlyJson",
-    "extract-api": "betools extract-api --entry=ExpressServer",
+    "docs": "betools docs --json=../../generated-docs/core/express-server/file.json --tsIndexFile=./express-server.ts --onlyJson",
+    "extract-api": "betools extract-api --entry=express-server",
     "lint": "eslint \"./src/**/*.ts\" 1>&2",
     "test": "mocha",
     "cover": "nyc npm -s test"
