+++ core/mobile/package.json
@@ -10,8 +10,8 @@
     "build": "npm run -s build:cjs",
     "build:cjs": "tsc 1>&2 --outDir lib/cjs",
     "clean": "rimraf -g lib .rush/temp/package-deps*.json",
-    "docs": "betools docs --json=../../generated-docs/core/core-mobile/file.json --tsIndexFile=./__DOC_ONLY_MOBILE__.ts --onlyJson",
-    "extract-api": "betools extract-api --entry=__DOC_ONLY_MOBILE__",
+    "docs": "betools docs --json=../../generated-docs/core/core-mobile/file.json --tsIndexFile=./core-mobile.ts --onlyJson",
+    "extract-api": "betools extract-api --entry=core-mobile",
     "lint": "eslint \"./src/**/*.ts\" 1>&2",
     "test": "",
     "cover": ""
