+++ tools/perf-tools/package.json
@@ -2,8 +2,8 @@
   "name": "@itwin/perf-tools",
   "version": "5.0.0-dev.108",
   "description": "Tools for collecting and reporting performance data",
-  "main": "lib/cjs/index.js",
-  "typings": "lib/cjs/index",
+  "main": "lib/cjs/perf-tools.js",
+  "typings": "lib/cjs/perf-tools",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -15,7 +15,7 @@
     "build:cjs": "tsc 1>&2 --outDir lib/cjs",
     "clean": "rimraf -g lib .rush/temp/package-deps*.json",
     "lint": "eslint \"./reporter/**/*.ts\" 1>&2",
-    "extract-api": "betools extract-api --entry=Reporter",
+    "extract-api": "betools extract-api --entry=perf-tools",
     "test": "",
     "docs": "",
     "cover": ""
